+++ src/commands/admin/blacklist.js
@@ -3,8 +3,8 @@ const Database = require("../../handlers/DatabaseManager");
 const { errorEmbed } = require("../../util/embeds");
 
 module.exports = {
-    name: "whitelist",
-    descripiton: "Manage users whitelist.",
+    name: "blacklist",
+    descripiton: "Manage users blacklist.",
     acceptDirectMessages: null,
     permissions: [],
     botPermissions: [],
@@ -12,7 +12,7 @@ module.exports = {
     ownerOnly: false,
     adminOnly: false,
     blacklistAllowed: false,
-    whitelistAllowed: false,
+    whitelistAllowed: true,
     options: [],
     ephemeral: false,
     /**
@@ -22,7 +22,7 @@ module.exports = {
      * @param {Database} db 
      */
     execute: async (bot, interaction, db) => {
-        let list = await db.getValue("client", bot.user.id, "whitelist");
+        let list = await db.getValue("client", bot.user.id, "blacklist");
         let s = 0 // start
         let e = 9 // end for paginator index from 0 to 9, from s to e
 
@@ -30,35 +30,35 @@ module.exports = {
             return new ActionRowBuilder()
             .addComponents(
                 new ButtonBuilder()
-                .setCustomId("whitelist_page_back")
+                .setCustomId("blacklist_page_back")
                 .setDisabled(s == 0)
                 .setEmoji(bot.emojisList.leftarrow)
                 .setStyle(ButtonStyle.Primary)
             )
             .addComponents(
                 new ButtonBuilder()
-                .setCustomId("whitelist_user_add")
+                .setCustomId("blacklist_user_add")
                 .setDisabled(list.length == 25 || ((message && message.embeds[0].fields.length === 10) ?? false))
                 .setEmoji(bot.emojisList.join)
                 .setStyle(ButtonStyle.Success)
             )
             .addComponents(
                 new ButtonBuilder()
-                .setCustomId("whitelist_reset")
+                .setCustomId("blacklist_reset")
                 .setDisabled(list.length < 1 || ((message && message.embeds[0].fields.length === 10) ?? false))
                 .setEmoji(bot.emojisList.ben)
                 .setStyle(ButtonStyle.Danger)
             )
             .addComponents(
                 new ButtonBuilder()
-                .setCustomId("whitelist_user_remove")
+                .setCustomId("blacklist_user_remove")
                 .setDisabled(list.length < 1 || ((message && message.embeds[0].fields.length === 10) ?? false))
                 .setEmoji(bot.emojisList.leave)
                 .setStyle(ButtonStyle.Danger)
             )
             .addComponents(
                 new ButtonBuilder()
-                .setCustomId("whitelist_page_next")
+                .setCustomId("blacklist_page_next")
                 .setDisabled(list.length <= e)
                 .setEmoji(bot.emojisList.rightarrow)
                 .setStyle(ButtonStyle.Primary)
@@ -68,7 +68,7 @@ module.exports = {
         const embed = new EmbedBuilder()
         .setColor(bot.colors.true)
         .setFooter({"text": "Powered by Aunt Development.", "iconURL": bot.user.displayAvatarURL({extension: "png", forceStatic: false, size: 2048})})
-        .setDescription(`${bot.emojisList.chat} - **You're managing the white list.** - \`${list.length} entries.\``)
+        .setDescription(`${bot.emojisList.chat} - **You're managing the black list.** - \`${list.length} entries.\``)
 
         
         if(list.length < 1) {
@@ -92,28 +92,28 @@ module.exports = {
 
             if(!button_i.deferred) await button_i.deferUpdate()
 
-            if(["whitelist_page_back", "whitelist_page_next"].includes(button_i.customId)) {
-                button_i.customId === "whitelist_page_back" ? s -= 9 : s += 9
-                button_i.customId === "whitelist_page_back" ? e -= 9 : e += 9
+            if(["blacklist_page_back", "blacklist_page_next"].includes(button_i.customId)) {
+                button_i.customId === "blacklist_page_back" ? s -= 9 : s += 9
+                button_i.customId === "blacklist_page_back" ? e -= 9 : e += 9
 
                 message.embeds[0].fields[0] = {name: message.embeds[0].fields[0].name, value: list.map((id) => `${bot.emojisList.user} - <@${id}> - \`${id}\``).slice(s, e).join("\n"), inline: false}
                 return void await message.edit({components: [getComponents(bot, list, e, s)], embeds: [message.embeds[0]]})
             }
 
-            if(["whitelist_user_add", "whitelist_user_remove"].includes(button_i.customId)) {
+            if(["blacklist_user_add", "blacklist_user_remove"].includes(button_i.customId)) {
                 let embed_1 = new EmbedBuilder()
                 .setColor(bot.colors.true)
                 .setFooter({"text": "Powered by Aunt Development.", "iconURL": bot.user.displayAvatarURL({extension: "png", forceStatic: false, size: 2048})})
                 .setDescription(`${bot.emojisList.chat} - **Give us any id.**`)
                 let menu_1
-                if(button_i.customId === "whitelist_user_add") {
+                if(button_i.customId === "blacklist_user_add") {
                     menu_1 = new ActionRowBuilder()
                     .addComponents(
                         new UserSelectMenuBuilder()
                         .setCustomId("menu_add")
                         .setMaxValues(1)
                         .setMinValues(1)
-                        .setPlaceholder("Add a new user to the whitelist!")
+                        .setPlaceholder("Add a new user to the blacklist!")
                         .setDisabled(false)
                     )
                 } else {
@@ -123,7 +123,7 @@ module.exports = {
                         .setCustomId("menu_remove")
                         .setMaxValues(1)
                         .setMinValues(1)
-                        .setPlaceholder("Add a new user to the whitelist!")
+                        .setPlaceholder("Add a new user to the blacklist!")
                         .setDisabled(false)
                         .setOptions(list.map(user => 
                             new StringSelectMenuOptionBuilder()
@@ -149,37 +149,37 @@ module.exports = {
                     if(!button_i.deferred) await button_i.deferUpdate()
                     await replied_message.delete();
 
-                    if(button_i.customId === "whitelist_user_add") {
+                    if(button_i.customId === "blacklist_user_add") {
                         if(list.includes(menu_i.values[0])) return;
-                        await db.addValue("client", bot.user.id, "whitelist", menu_i.values[0])
+                        await db.addValue("client", bot.user.id, "blacklist", menu_i.values[0])
                     } else {
-                        await db.removeValue("client", bot.user.id, "whitelist", menu_i.values[0])
+                        await db.removeValue("client", bot.user.id, "blacklist", menu_i.values[0])
                     }
-                    list = await db.getValue("client", bot.user.id, "whitelist")
+                    list = await db.getValue("client", bot.user.id, "blacklist")
 
                     bot.users.cache.get(menu_i.values[0]).send({embeds: [
                         new EmbedBuilder()
                         .setColor(bot.colors.true)
                         .setFooter({"text": "Powered by Aunt Development.", "iconURL": bot.user.displayAvatarURL({extension: "png", forceStatic: false, size: 2048})})
-                        .setDescription(`${bot.emojisList.chat} - **You've been ${button_i.customId === "whitelist_user_add" ? `added in` : `removed from`} the white list.**`)
+                        .setDescription(`${bot.emojisList.chat} - **You've been ${button_i.customId === "blacklist_user_add" ? `added in` : `removed from`} the black list.**`)
                     ]})
 
                     message.embeds[0].fields[0] = list.length < 1 ? {name: message.embeds[0].fields[0].name, value: `${bot.emojisList.user} - There's no user on the list.`, inline: false} : {name: message.embeds[0].fields[0].name, value: list.map((id) => `${bot.emojisList.user} - <@${id}> - \`${id}\``).slice(s, e).join("\n"), inline: false}
-                    message.embeds[0].fields[message.embeds[0].fields.length] = {name: `<t:${Math.round(Date.now() / 1000)}:R>`, value: `${bot.emojisList.chat} - <@${button_i.user.id}> - <@${menu_i.values[0]}> has been ${button_i.customId === "whitelist_user_add" ? `added in` : `removed from`} the white list.`, inline: false}
-                    message.embeds[0].description = `${bot.emojisList.chat} - **You're managing the white list.** - \`${list.length} entries.\``
+                    message.embeds[0].fields[message.embeds[0].fields.length] = {name: `<t:${Math.round(Date.now() / 1000)}:R>`, value: `${bot.emojisList.chat} - <@${button_i.user.id}> - <@${menu_i.values[0]}> has been ${button_i.customId === "blacklist_user_add" ? `added in` : `removed from`} the black list.`, inline: false}
+                    message.embeds[0].description = `${bot.emojisList.chat} - **You're managing the black list.** - \`${list.length} entries.\``
                     return void await message.edit({components: [getComponents(bot, list, e, s, message)], embeds: [message.embeds[0]]})
                 })
             }
 
-            if(button_i.customId === "whitelist_reset") {
-                await db.setValue("client", bot.user.id, "whitelist", []);
+            if(button_i.customId === "blacklist_reset") {
+                await db.setValue("client", bot.user.id, "blacklist", []);
                 list = []
                 s = 0
                 e = 9
                 
                 message.embeds[0].fields[0] = {name: message.embeds[0].fields[0].name, value: `${bot.emojisList.user} - There's no user on the list.`, inline: false}
                 message.embeds[0].fields[message.embeds[0].fields.length] = {name: `<t:${Math.round(Date.now() / 1000)}:R>`, value: `${bot.emojisList.chat} - <@${button_i.user.id}> - The list has been reset.`, inline: false}
-                message.embeds[0].description = `${bot.emojisList.chat} - **You're managing the white list.** - \`${list.length} entries.\``
+                message.embeds[0].description = `${bot.emojisList.chat} - **You're managing the black list.** - \`${list.length} entries.\``
                 return void await message.edit({components: [getComponents(bot, list, e, s, message)], embeds: [message.embeds[0]]})
             }
         })