+++ src/commands/util/ping.js
@@ -23,18 +23,13 @@ module.exports = {
     execute: async (bot, interaction, db) => {
         const embed = new EmbedBuilder()
         .setColor(bot.colors.true)
-        .setFooter({
-            "text": "Powered by Aunt Development.",
-            "iconURL": bot.user.displayAvatarURL({extension: "png", forceStatic: false, size: 2048})
-        })
+        .setFooter({"text": "Powered by Aunt Development.", "iconURL": bot.user.displayAvatarURL({extension: "png", forceStatic: false, size: 2048})})
 
         let message = await interaction.editReply({embeds: [embed.setDescription(`${bot.emojisList.chat} - **Fetching latencies...**`)]})
-        await message.edit({
-            embeds: [
-                embed
-                .setDescription(`${bot.emojisList.chat} - **Latencies fetched!**`)
-                .addFields({name: `<t:${Math.round(Date.now() / 1000)}:R>`, value: `${bot.emojisList.stats} - **Passerelle discord** - \`${bot.ws.ping}ms\`\n${bot.emojisList.support} - **Réponses** - \`${Math.round(Date.now() - message.createdTimestamp)}ms\``, inline: false})
-            ]
-        })
+        await message.edit({embeds: [
+            embed
+            .setDescription(`${bot.emojisList.chat} - **Latencies fetched!**`)
+            .addFields({name: `<t:${Math.round(Date.now() / 1000)}:R>`, value: `${bot.emojisList.stats} - **Passerelle discord** - \`${bot.ws.ping}ms\`\n${bot.emojisList.support} - **Réponses** - \`${Math.round(Date.now() - message.createdTimestamp)}ms\``, inline: false})
+        ]})
     }
 }