+++ CHANGELOG.md
@@ -2,9 +2,21 @@
  🦫 〢 Changelog
 </h1>
 
+# <a id="menu"></a>🍃 〢 Menu
+
+- [Version BÊTA](#v0)
+    - [0.5](#v0-5)
+    - [0.1](#v0-1)
+
 ---
-# <a id="menu"></a>🍃 〢 2024
-### <a id="setup"></a> 📁 〢 Juillet
-- Slash command handler
-- Event handler
-- Ping command
+# <a id="v0"></a>🍃 〢 2024
+### <a id="v0-5"></a> 📁 〢 V0.5.0
+- Created: Database manager.
+- Added: Whitelist & Blacklist systems.
+- Fixed: Slash commands handler - (Uknow interaction error).
+- Added: Blacklist & Whitelist to the handler of commands.
+
+### <a id="v0-1"></a> 📁 〢 V
+- Created: Slash commands handler.
+- Created: Event handler.
+- Added: Ping command.